Add stats to group
